Fixed errors found during run through of all datacrunchers
